# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Business Steward — Backend API
  description: |
    The HTTP API backing **My Business Steward** (Restaurant Margin Copilot) —
    a daily-close and margin-alert copilot for an independent restaurant.

    This spec documents every route actually registered in
    `backend/cmd/server/main.go`. It was written by reading the Go handler
    source for each route (`backend/internal/httpapi/*.go`,
    `backend/internal/badges/badges.go`) and cross-checked against `curl`
    calls to a live instance of this backend seeded from
    `backend/data/live/` — request/response shapes are not guessed from
    route names.

    ### The deterministic / probabilistic split
    Every endpoint except `POST /api/ask` and `POST /api/business-insight`
    is **pure deterministic REST**: plain reads or writes against
    Postgres-persisted reconciliation output, with no model anywhere in
    the request path. `POST /api/ask` calls an LLM (Claude Sonnet 5 for an
    ambiguity gate, Claude Sonnet 5 to narrate an already-computed number,
    Claude Haiku 4.5 for the paraphrase-match cache classifier) — and even
    there, the model never computes a figure; it only classifies the
    question and narrates a result the Go reconciliation engine already
    produced. `POST /api/business-insight` (specs/009) is the one endpoint
    that returns openly probabilistic content — general business advice —
    and it is opt-in by construction: its trigger (`business_insight` on
...

...
servers:
...
security: []
...
paths:
...
  /api/badges:
    get:
      tags: [Badges]
      summary: Get earned badges and points balance
      description: |
        Evaluates all five badge codes across four categories
        (Reconciliation, Growth, Engagement, Campaign-Creation) at read
        time from already-persisted data — nothing here is stored state.

        `?start`/`?end` are **optional** and, per
        `backend/internal/badges/badges.go`'s `RegisterBadgeHandler` doc
        comment, scope **Reconciliation-category badges only**
        (Clean Close / Discrepancy Catcher). Growth, Engagement, and
        Campaign-Creation badges are always evaluated over **all**
        persisted data regardless of this query, the same way
        `GET /api/promotions` is a full listing rather than a period query.
      operationId: getBadges
      parameters:
        - name: start
          in: query
          required: false
          description: Inclusive start date (YYYY-MM-DD). Scopes Reconciliation-category badges only. If given, does not require `end`.
          schema: { type: string, format: date, example: "2026-08-01" }
        - name: end
          in: query
          required: false
          description: Inclusive end date (YYYY-MM-DD). Scopes Reconciliation-category badges only.
          schema: { type: string, format: date, example: "2026-08-14" }
      responses:
        "200":
          description: Badges and points, successfully evaluated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BadgesResponse" }
              example:
                badges:
                  - date: "2026-08-01"
                    code: clean_close
                    name: Clean Close
                    category: reconciliation
                  - date: "2026-08-03"
                    code: discrepancy_catcher
                    name: Discrepancy Catcher
                    category: reconciliation
                    discrepancy_flag_types: [duplicate_order_removed]
                  - date: "2026-08-07"
                    code: growth
                    name: Growth
                    category: growth
                    campaign_id: IFOOD-CAMP-BOOST01
                points:
                  total: 200
                  breakdown:
                    - { code: clean_close, name: Clean Close, count: 12, points_each: 10, points: 120 }
                    - { code: discrepancy_catcher, name: Discrepancy Catcher, count: 2, points_each: 25, points: 50 }
                    - { code: growth, name: Growth, count: 2, points_each: 15, points: 30 }
        "400":
          description: A supplied `start`/`end` is not a valid YYYY-MM-DD date.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example: { error: invalid_period, detail: "parsing time \"not-a-date\" as \"2006-01-02\": cannot parse \"not-a-date\" as \"2006\"" }
        "405":
          description: Any method other than GET.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example: { error: method_not_allowed, detail: "only GET is supported" }
        "500":
          description: A Postgres query failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ------------------------------------------------------------------
  /api/reconciliation:
    get:
      tags: [Reconciliation]
      summary: List daily reconciliation results
      description: |
        Returns one `DailySummaryResult` per persisted `DailyReconciliation`
        row in range. A calendar day with no reconciliation is simply
        **absent** from `days`, never present with zeroed figures — a
        missing day and a zero day are different facts.

        `start`/`end` are optional and independent (either may be given
        without the other); omitting both serves everything persisted.
        This is the exact rendering `GET /api/badges`'s period query and a
        chat answer about the same day share — one computation, one
        formatting, everywhere.
      operationId: getReconciliation
      parameters:
        - name: start
          in: query
          required: false
          schema: { type: string, format: date, example: "2026-08-01" }
        - name: end
          in: query
          required: false
          schema: { type: string, format: date, example: "2026-08-02" }
      responses:
        "200":
          description: The reconciliation days in range (possibly empty).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReconciliationResponse" }
              example:
                start: "2026-08-01"
                end: "2026-08-02"
                days:
                  - date: "2026-08-01"
                    gross_sales_by_source: { ifood: "69.50", just_eat_takeaway: "76.25", pos: "248.75" }
                    total_delivery_gross_sales: "145.75"
                    commissions: "31.24"
                    refunds: "0.00"
                    input_costs: "320.00"
                    margin: "43.26"
                    discrepancy_flags: []
                    source_row_refs:
                      - { file: "/data/live/delivery_platform_export.csv", row: 2 }
                      - { file: "/data/live/pos_export.csv", row: 2 }
                      - { file: "/data/live/supplier_cost_sheet.csv", row: 2 }
                  - date: "2026-08-02"
                    gross_sales_by_source: { ifood: "72.50", just_eat_takeaway: "81.75", pos: "223.75" }
                    total_delivery_gross_sales: "154.25"
                    commissions: "25.09"
                    refunds: "34.50"
                    input_costs: "545.50"
                    margin: "-227.09"
                    discrepancy_flags: []
                    source_row_refs:
                      - { file: "/data/live/delivery_platform_export.csv", row: 6 }
        "400":
          description: end is before start, or a date is malformed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example: { error: invalid_period, detail: "end 2026-08-01 is before start 2026-08-05" }
        "405":
          description: Any method other than GET.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: A Postgres query failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ------------------------------------------------------------------
  /api/promotions:
    get:
      tags: [Promotions]
      summary: List all promotion ROI records
      description: |
        Every persisted `PromotionRoiRecord` — both pipeline-ingested
        campaigns and owner-created ones (see `POST /api/promotions`) — in
        the same shape `get_promotion_roi` returns to the chat surface.
        This is a full listing, not a period query, matching
        `GET /api/badges`'s Growth/Engagement/Campaign-Creation behavior:
        "which campaigns paid for themselves" is not a question a default
        date window can answer without silently omitting real records.

        `roi`, `attributed_incremental_orders`, and
        `attributed_incremental_revenue` are `null` together, with
        `reason: "attribution_unavailable"` set, exactly when the
        deterministic layer could not attribute incremental orders to the
        campaign (never a fabricated zero).
      operationId: listPromotions
      responses:
        "200":
          description: All promotion records.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PromotionsResponse" }
              example:
                promotions:
                  - platform: iFood
                    campaign_id: IFOOD-CAMP-BOOST01
                    period: { start: "2026-08-01", end: "2026-08-07" }
                    spend: "180.00"
                    attributed_incremental_orders: 6
                    attributed_incremental_revenue: "214.00"
                    roi: "34.00"
                    flagged_negative: false
                    source_row_refs:
                      - { file: "data/live/promotion_ad_spend_export.csv", row: 2 }
                      - { file: "data/live/delivery_platform_export.csv", row: 2 }
                    origin: ingested
                  - platform: Just Eat Takeaway
                    campaign_id: JET-CAMP-LUNCHFIX
                    period: { start: "2026-08-04", end: "2026-08-10" }
                    spend: "220.00"
                    attributed_incremental_orders: 2
                    attributed_incremental_revenue: "55.00"
                    roi: "-165.00"
                    flagged_negative: true
                    source_row_refs:
                      - { file: "data/live/promotion_ad_spend_export.csv", row: 3 }
                    origin: ingested
                  - platform: iFood
                    campaign_id: IFOOD-CAMP-WEEKEND
                    period: { start: "2026-08-08", end: "2026-08-09" }
                    spend: "95.00"
                    attributed_incremental_orders: null
                    attributed_incremental_revenue: null
                    roi: null
                    reason: attribution_unavailable
                    flagged_negative: false
                    source_row_refs:
                      - { file: "data/live/promotion_ad_spend_export.csv", row: 4 }
                    origin: ingested
        "405":
          description: Any method other than GET (the handler this route dispatches to on a non-POST method).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: A Postgres query failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      tags: [Promotions]
      summary: Log a new owner-created promotion record
      description: |
        The owner logging a promotion directly in the app (spec
        002-badge-expansion, User Story 3) — not an ingested pipeline
        record, and never routed through the model's MCP tool boundary.

        If `replaces` is set, the referenced `campaign_id`'s
        `flagged_negative` fact is **re-verified server-side, against the
        live database**, at submission time — a client cannot earn the
        Campaign-Creation badge by lying about what its own UI displayed.
        A `replaces` claim against a campaign that is not currently
        flagged negative is refused with `422 replaces_not_flagged_negative`;
        the promotion can still be logged by omitting `replaces`.
      operationId: createPromotion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePromotionRequest" }
            example:
              platform: iFood
              campaign_id: IFOOD-CAMP-AUTUMN
              period: { start: "2026-08-15", end: "2026-08-21" }
              spend: "120.00"
              replaces: JET-CAMP-LUNCHFIX
      responses:
        "201":
          description: The promotion was created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatePromotionResponse" }
              example:
                promotion:
                  platform: iFood
                  campaign_id: IFOOD-CAMP-AUTUMN
                  period: { start: "2026-08-15", end: "2026-08-21" }
                  spend: "120.00"
                  attributed_incremental_orders: null
                  attributed_incremental_revenue: null
                  roi: null
                  flagged_negative: false
                  source_row_refs: []
                  origin: owner_created
                  replaces_campaign_id: JET-CAMP-LUNCHFIX
                earned_campaign_creation_badge: true
        "400":
          description: Missing required field(s), a malformed period, or a malformed/negative spend amount.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example: { error: invalid_input, detail: "missing required field(s): platform, campaign_id, period.start/period.end, spend" }
        "405":
          description: Any method other than POST.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: A promotion for this exact platform, campaign_id, and period already exists (unique constraint).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example: { error: already_exists, detail: "a promotion for platform \"iFood\", campaign_id \"IFOOD-CAMP-AUTUMN\", and this exact period already exists" }
        "422":
          description: A `replaces` claim was made against a campaign_id that is not currently flagged negative-ROI.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example: { error: replaces_not_flagged_negative, detail: "campaign_id \"IFOOD-CAMP-BOOST01\" is not currently flagged negative-ROI by list_negative_roi_promotions — refusing the replaces claim. The promotion can still be logged without a replaces reference." }
        "500":
          description: A Postgres query failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ------------------------------------------------------------------
...
  /api/platforms/trend:
    get:
      tags: [Platforms]
      summary: Effective-rate trend across the trailing calendar months
      description: |
        Spec 008 FR-007's Platforms effective-rate trend — the same
        `compare_platform_economics` computation `GET /api/platforms`
        exposes, run server-side across the trailing 6 calendar months
        ending on the real persisted data's own max date. A bounded, fully
        deterministic aggregation (no model call): a genuinely new small
        endpoint rather than several sequential frontend calls, since
        fetching up to 6 months one at a time would be chattier than this
        project's own "2-3 sequential calls" guideline for when to prefer
        assembling from existing endpoints instead.

        A calendar month with no persisted reconciliation for any day in
        it is silently **omitted** from `periods` — never zero-padded or
        fabricated (spec 008 FR-013). The most recent entry's `end` is
        truncated to the real data's max date rather than padded to the
        end of that calendar month.
      operationId: getPlatformsTrend
      responses:
        "200":
          description: 0-6 real, non-refused trailing months, oldest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlatformsTrendResponse" }
              example:
                periods:
                  - month: "2026-08"
                    result:
                      period: { start: "2026-08-01", end: "2026-08-14" }
                      days_included: 14
                      platforms:
                        - source: ifood
                          display_name: iFood
                          gross_sales: "838.00"
                          effective_rate: "22.06%"
        "405":
          description: Any method other than GET.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: A Postgres query failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ------------------------------------------------------------------
  /api/usage:
    post:
      tags: [Usage]
      summary: Record an app-open usage ping
      description: |
        The real app-open ping backing Engagement badge evaluation (spec
        002-badge-expansion). Deliberately the thinnest possible handler —
        no-double-counting-within-a-calendar-day is enforced at the
        database layer (a unique index on a generated column), not here.
        Fires once per app load regardless of whether today was already
        recorded; the frontend does not need to branch on the response.
      operationId: recordUsage
      responses:
        "200":
          description: The ping was recorded (or today was already on file).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecordUsageResponse" }
              examples:
                firstPingToday:
                  summary: First ping today
                  value: { recorded: true }
                repeatPingToday:
                  summary: A later ping the same day
                  value: { recorded: false }
        "405":
          description: Any method other than POST.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: A Postgres query failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ------------------------------------------------------------------
...
components:
  schemas:
...
    BadgesResponse:
      type: object
      properties:
        badges:
          type: array
          items: { $ref: "#/components/schemas/Badge" }
        points: { $ref: "#/components/schemas/Points" }
      required: [badges, points]

    # --- reconciliation ---------------------------------------------------
...
    PromotionsResponse:
      type: object
      properties:
        promotions:
          type: array
          items: { $ref: "#/components/schemas/PromotionRoiView" }
      required: [promotions]
...